Extract numerical minimization from test_symmetric
Testing using scipy.minimize should be used more extensively, so the
procrustes/test/common.py was added. The current functions only work
for symmetirc matrices, so they should be extended.
